竞猜排行榜数据接入；
git-svn-id: svn://ci.ftxgame.local:8199/repos/ArmChairQuatebackDoc/sourcecode/arm_chair_quaterback@646694 11eda05a-b276-4d74-a309-1d70aa31e1c1

+++ lib/common/entities/rank_list_entity.dart
@@ -34,7 +34,7 @@ class RankListMyRank {
   int? rank;
   int? updateTime;
   int? cycle;
-  int? win;
+  double? win;
 
   RankListMyRank();
